Simplified plot

+++ cpp/main_application/plot_objects/plot2d/plot2d.h
@@ -12,7 +12,7 @@
 class Plot2D : public PlotObjectBase
 {
 private:
-    uint8_t* points_ptr_;
+    float* points_ptr_;
 
     void findMinMax() override;
